Change Double for integer in Starbuzz

+++ src/main/java/starbuzz/Beverage.java
@@ -17,5 +17,5 @@ public abstract class Beverage {
 		this.size = size;
 	}
  
-	public abstract Double cost();
+	public abstract Integer cost();
 }
